+++ NEWS.md
@@ -1,5 +1,7 @@
 # milQuant 1.0.1 _tba_
 * the connection settings (name/pw) are reusable (with milQuant-dist)
+* improved plot legends
+* fixed period order in pottery & quant B plots
 
 # milQuant 1.0.0 _2023-02-21_
 * first "release" (version number changed to be same as the standalone version)